Add /account/bids_placed
This replaces the previous /my-bids controller action with some new and
slightly different functionality, similar to the bids display for users
in the admin.
