+++ app/assets/stylesheets/components/_profile.scss
@@ -1,7 +1,6 @@
 .profile-form {
   @include span-columns(8);
   @include shift(2);
-  margin-top: 1em;
   padding-left: 1em;
   padding-right: 1em;
