+++ app/assets/stylesheets/components/_users-header.scss
@@ -29,3 +29,7 @@
     box-shadow: none;
   }
 }
+
+.user-body-wrapper {
+  margin-top: 1em;
+}
