+++ app/controllers/users_controller.rb
@@ -15,4 +15,8 @@ class UsersController < ApplicationController
       render :edit
     end
   end
+
+  def bids_placed
+    @view_model = AccountBidsPlacedViewModel.new(current_user: current_user)
+  end
 end
