+++ app/view_models/account_bid_list_item.rb
