+++ app/view_models/admin/user_auction_view_model.rb
@@ -1,42 +1,6 @@
-class Admin::UserAuctionViewModel
-  attr_reader :auction, :user
-
-  DATE_FORMAT = '%m/%d/%Y'.freeze
-  NA_RESPONSE_STRING = '-'.freeze
-
-  def initialize(auction, user)
-    @auction = auction
-    @user = user
-  end
-
-  def title
-    auction.title
-  end
-
-  def id
-    auction.id
-  end
-
-  def bidding_status
-    BiddingStatusPresenterFactory.new(auction).create.label
-  end
-
+class Admin::UserAuctionViewModel < UserAuctionViewModel
   def skills
     auction.sorted_skill_names.join(', ')
-  end
-
-  def user_bid_count
-    user_bids.count
-  end
-
-  def user_won_label
-    if user_won?
-      'Yes'
-    elsif auction_over?
-      'No'
-    else
-      NA_RESPONSE_STRING
-    end
   end
 
   def accepted_label
@@ -51,23 +15,7 @@ class Admin::UserAuctionViewModel
 
   private
 
-  def user_won?
-    auction_over? && WinningBid.new(auction).find.bidder == user
-  end
-
-  def auction_over?
-    BiddingStatus.new(auction).over?
-  end
-
   def auction_accepted?
     auction.accepted?
-  end
-
-  def auction_rules
-    RulesFactory.new(auction).create
-  end
-
-  def user_bids
-    auction.bids.where(bidder: user).sort_by(&:amount)
   end
 end
