+++ app/views/components/_user_nav_drawer.html.erb
@@ -3,7 +3,7 @@
   <%= link_to 'My profile', profile_path %>
 </li>
 <li class="nav-link">
-  <%= link_to 'My bids', bids_path %>
+  <%= link_to 'My bids', bids_placed_path %>
 </li>
 <li class="nav-link">
   <%= link_to 'Logout', logout_path %>
