+++ app/views/users/_bids_table.html.erb
@@ -1,0 +1,22 @@
+<table class="striped-table">
+  <thead>
+    <tr>
+      <th scope="col"><%= t 'labels.vendor.account.bids_placed.columns.auction' %></th>
+      <th scope="col"><%= t 'labels.vendor.account.bids_placed.columns.status' %></th>
+      <th scope="col"><%= t 'labels.vendor.account.bids_placed.columns.number_bids' %></th>
+      <th scope="col"><%= t 'labels.vendor.account.bids_placed.columns.lowest_bid' %></th>
+      <th scope="col"><%= t 'labels.vendor.account.bids_placed.columns.won' %></th>
+    </tr>
+  </thead>
+  <tbody>
+    <% view_model.auctions.each do |auction| %>
+    <tr>
+      <td><%= link_to(auction.title, auction_path(auction.id)) %></td>
+      <td><%= auction.bidding_status %></td>
+      <td><%= auction.user_bid_count %></td>
+      <td><%= auction.user_lowest_bid %></td>
+      <td><%= auction.user_won_label %></td>
+    </tr>
+    <% end %>
+  </tbody>
+</table>
