+++ app/views/users/_bids_table_none.html.erb
@@ -1,0 +1,2 @@
+<%= t('labels.vendor.account.bids_placed.no_bids_html',
+    index_url: Url.new(link_text: 'current and upcoming auctions', path_name: 'root')) %>
