+++ app/views/users/edit.html.erb
@@ -8,7 +8,7 @@
     </div>
   </div>
   <div class="usa-grid">
-    <div class="profile-form">
+    <div class="user-body-wrapper profile-form">
       <%= simple_form_for @view_model.record, method: :put do |f| %>
         <h3>Contact information</h3>
         <%= f.input :name %>
