+++ features/step_definitions/page_content_steps.rb
@@ -59,3 +59,7 @@ end
 Then(/^the "([^"]+)" subnav should be selected$/) do |text|
   page.find('a.nav-auction.active', text: text)
 end
+
+When(/^I click on the "([^"]+)" subnav$/) do |text|
+  click_link(text)
+end
