+++ features/vendor_views_profile.feature
@@ -9,3 +9,22 @@ Feature: Vendor views profile
     Then I should see a page header labeled "My Account"
     And the "Profile" subnav should be selected
     And I should see a profile form with my info
+
+  Scenario: Vendor who has not bid navigates to bid history page
+    Given I am an authenticated vendor
+    When I visit my profile page
+    Then I should see a page header labeled "My Account"
+
+    When I click on the "Bids placed" subnav
+    Then the "Bids placed" subnav should be selected
+    And I should see I have placed no bids
+
+  Scenario: Vendor who has placed bids navigates to the bid history page
+    Given I am an authenticated vendor
+    And there is an open auction
+    And I have placed a bid
+    When I visit my bid history page
+    Then I should see a page header labeled "My Account"
+    And the "Bids placed" subnav should be selected
+    And I should see my bid history
+    And I should not see bids from other users
