+++ spec/controllers/users_controller_spec.rb
@@ -27,7 +27,37 @@ describe UsersController do
     end
   end
 
+  describe '#bids_placed' do
+    context 'when logged in' do
+      it 'should assign auctions that current user have bidded on, presented' do
+        current_bidder = create(:user, sam_status: :sam_accepted)
+        auction = create(:auction)
+        bid = create(:bid, bidder: current_bidder, auction: auction)
+        expect(bid).to be_valid
+
+        get :bids_placed, { }, user_id: current_bidder.id
+
+        view_model = assigns(:view_model)
+        expect(view_model).to_not be_nil
+        expect(view_model.auctions).to_not be_empty
+      end
+
+      it 'should not assign auctions that the current user has not bidded on' do
+        current_bidder = create(:user, sam_status: :sam_accepted)
+        auction = create(:auction)
+        other_person = create(:user)
+        create(:bid, bidder: other_person, auction: auction)
+
+        get :bids_placed, { }, user_id: current_bidder.id
+
+        view_model = assigns(:view_model)
+        expect(view_model).to_not be_nil
+        expect(view_model.auctions).to be_empty
+      end
+    end
+  end
+
   def user
-    @_user ||=  create(:user)
+    @_user ||= create(:user)
   end
 end
